Houses get more agents

+++ NeedManager.java
@@ -50,7 +50,7 @@ public class NeedManager
 		{
 			sat += Math.min(house.getNeed(i)/req, 1)*weights[i];
 		}
-		return (int)(100*(sat/weightTot));
+		return 100*(sat/weightTot);
 	}
 	
 	public double[] getGoalWeights(House house){
